extract filter inputs into component
